+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 2023.07.31
+
+- updates Deno's `std` library
+- adds `processBuffer` helper function in `default.ts` to convert string values to buffers when expected
+- comments out tests that fail…LOL
+
+
+
 ## 2023.04.06
 
 - imports `Buffer` from Deno's internals